Started working on texture and sprite code
The drawing/rendering node code had to be reworked a bit. A display
object is now passed when rendering states. This is better than having
all of the states store a reference to the display - which would created
extra memory overhead and problems for when the display is disposed.
More work needs to be done to solidify this process.