Renaming to make things clearer
Renamed 'ID' to 'Value' and 'Value' to 'Strength'. ID should correspond
to a number that is assigned to a specific type/value combination. Value
is a numerical value associated with an input type (keyboard key code,
mouse button, etc.). Strength is how much of the input there is
(joystick distance).